refactored example folder to better fit detox testing
