+++ example/src/modals/SlowModal.tsx
@@ -1,0 +1,24 @@
+import React from 'react';
+
+// @ts-ignore
+import Modal from 'react-native-modal';
+import ModalBaseScene from '../utils/ModalBaseScene';
+import DefaultModalContent from '../utils/DefaultModalContent';
+
+class SlowModal extends ModalBaseScene {
+  renderModal(): React.ReactElement<any> {
+    return (
+      <Modal
+        testID={'modal'}
+        isVisible={this.isVisible()}
+        animationInTiming={1000}
+        animationOutTiming={1000}
+        backdropTransitionInTiming={800}
+        backdropTransitionOutTiming={800}>
+        <DefaultModalContent onPress={this.close} />
+      </Modal>
+    );
+  }
+}
+
+export default SlowModal;
