+++ package.json
@@ -48,7 +48,8 @@
   },
   "dependencies": {
     "prop-types": "^15.6.2",
-    "react-native-animatable": "1.3.3"
+    "react-native-animatable": "1.3.3",
+    "react-native-screens": "^1.0.0-alpha.23"
   },
   "devDependencies": {
     "@babel/core": "^7.5.5",
